slippery-slope/solution/requirements.txt
attrs==25.4.0
idna==3.11
ldn==0.0.15
outcome==1.3.0.post0
pycryptodome==3.23.0
python-netlink==0.0.14
python-pcapng==2.1.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.32.0